Clean up the RST doctree.
This change addresses several issues:
* Several docs are included multiple times.
* Several autodoc directives are duplicated.
* Subpackages do not have effective base module documentation targets.

To deduplicate documentation, the doctree is simply flattened a bit,
though section delimiters are adjusted to better represent the
document hierarchy.

Since there is no docstring for most packages (`__index__.py` files),
the `automodule` directive didn't do anything, and is replaced
with a static module directive in the rst document.

Ref #375